+++ app/App.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { StackNavigator } from 'react-navigation'
+import Home from './components/Home'
+
+const App = StackNavigator({
+  Home: {screen: Home},
+  // NewChore: {screen: NewChore},
+  // Group: {screen: Group}
+})
+
+export default App
