+++ package.json
@@ -21,7 +21,8 @@
   "dependencies": {
     "expo": "^16.0.0",
     "react": "16.0.0-alpha.6",
-    "react-native": "^0.43.4"
+    "react-native": "^0.43.4",
+    "react-navigation": "^1.0.0-beta.9"
   },
   "license": "MIT",
   "author": "Daniel Fattal, Jason Miguel, Jeff Kandel",
